+++ packages/tools/src/index.ts
@@ -1,6 +1,7 @@
 export * from '@/bn'
 export * from '@/duration'
 export * from '@/enums'
+export * from '@/errors'
 export * from '@/event-emitter'
 export * from '@/time'
 export * from '@/types'