+++ CHANGELOG.md
@@ -4,9 +4,31 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0-rc.11] - 2023-05-16
+### Added
+- `@distributedlab/w3p`
+  - EVM fallback provider
+  - readme details about custom provider implementation
+  - possibility to set details for many supported chains
+  - `wrapExternalProvider` method for EVM providers
+- `@distributedlab/tools` - `RuntimeError` class
+
+### Changed
+- `@distributedlab/w3p`
+  - `multiple-providers` examples about adding details for supported chains
+  - `SwitchChain` method of `ProviderBase` interface now optional
+  - errors classes
+  - `handleEthError` method
+  - evm `addChain` method
+  - `BaseEVMProvider` methods now ain't need try catch anymore
+
+### Removed
+- `@distributedlab/w3p` - `RuntimeError` class
+
+
 ## [0.2.0-rc.10] - 2023-05-11
 ### Changed
-- `@distributeedlab/w3p` - types, to be more flexible with external provider proxy constructors
+- `@distributedlab/w3p` - types, to be more flexible with external provider proxy constructors
 
 ## [0.2.0-rc.9] - 2023-05-02
 ### Fixed
@@ -177,7 +199,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [old repo]: https://github.com/distributed-lab/web-kit-old
 
-[Unreleased]: https://github.com/distributed-lab/web-kit/compare/0.2.0-rc.10...HEAD
+[Unreleased]: https://github.com/distributed-lab/web-kit/compare/0.2.0-rc.11...HEAD
+[0.2.0-rc.11]: https://github.com/distributed-lab/web-kit/compare/0.2.0-rc.10...0.2.0-rc.11
 [0.2.0-rc.10]: https://github.com/distributed-lab/web-kit/compare/0.2.0-rc.9...0.2.0-rc.10
 [0.2.0-rc.9]: https://github.com/distributed-lab/web-kit/compare/0.2.0-rc.8...0.2.0-rc.9
 [0.2.0-rc.8]: https://github.com/distributed-lab/web-kit/compare/0.2.0-rc.7...0.2.0-rc.8